+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Renamed `ftp-enforce-parens-in-return-single-element-tuple` to `ftp-disallow-parens-in-return-single-element-tuple`
 - Support type aliases in FTP077 and FTP104
+- Support `typing.Union` in FTP077 and FTP104
 
 ## 25.3.25.1